requirements.txt
aiogram>=3.19.0
aiohttp~=3.11.14
APScheduler>=3.11.0
aiosqlite>=0.21.0
fastapi>=0.115.0
httpx>=0.25.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
PyJWT>=2.8.0
python-dotenv>=1.0.1
pywebpush>=2.0.0
qrcode[pil]
requests~=2.32.3
remnawave
remnawave_api
uvicorn[standard]>=0.30.0
yookassa~=3.5.0
